requirement.txt
fastapi
uvicorn
pydantic
google-generativeai
streamlit
requests
faiss-cpu
langchain_community
sentence-transformers